refactor(usage-record): neutralize source naming
