+++ turbo/apps/api/src/signals/routes/__tests__/helpers/api-bdd-billing-media.ts
@@ -28,9 +28,9 @@ import { zeroImageIoGenerateContract } from "@okouai/api-contracts/contracts/zer
 import { mapsContract } from "@okouai/api-contracts/contracts/maps";
 import { zeroUsageMembersContract } from "@okouai/api-contracts/contracts/zero-usage";
 import {
-  zeroUsageRecordContract,
+  usageRecordContract,
   type UsageRecordRange,
-} from "@okouai/api-contracts/contracts/zero-usage-record";
+} from "@okouai/api-contracts/contracts/usage-record";
 import { zeroVideoIoGenerateContract } from "@okouai/api-contracts/contracts/zero-video-io-generate";
 import { voiceIoQuotaContract } from "@okouai/api-contracts/contracts/voice-io-quota";
 import { voiceIoSpeechContract } from "@okouai/api-contracts/contracts/voice-io-speech";
@@ -64,7 +64,7 @@ import { zeroFeatureSwitchesRoutes } from "../../zero-feature-switches";
 import { zeroImageIoGenerateRoutes } from "../../zero-image-io-generate";
 import { mapsRoutes } from "../../maps";
 import { zeroUsageMembersRoutes } from "../../zero-usage-members";
-import { zeroUsageRecordRoutes } from "../../zero-usage-record";
+import { usageRecordRoutes } from "../../usage-record";
 import { zeroVideoIoGenerateRoutes } from "../../zero-video-io-generate";
 import { voiceIoQuotaRoutes } from "../../voice-io-quota";
 import { voiceIoSpeechRoutes } from "../../voice-io-speech";
@@ -522,8 +522,8 @@ export function createBillingMediaApi(context: TestContext) {
     },
 
     async readUsageRecord(actor: ApiTestUser) {
-      const client = setupApp({ context, routes: zeroUsageRecordRoutes })(
-        zeroUsageRecordContract,
+      const client = setupApp({ context, routes: usageRecordRoutes })(
+        usageRecordContract,
       );
       return await accept(
         client.get({
